added user regiteration

+++ todolist/views.py
@@ -4,18 +4,29 @@ from django.contrib import messages
 from django.core.paginator import Paginator
 from .models import TaskList
 from .forms import TaskForm
+from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
 
 
+def index(request):
+    print('inside index')
+    return render(request, 'index.html', {})
+
+
+@login_required
 def todolist(request):
     if request.method == "POST":
         print(request.method, request.POST)
         form = TaskForm(request.POST or None)
         if form.is_valid():
-            form.save()
+            instance = form.save(commit=False)
+            instance.user = request.user
+            instance.save()
         messages.success(request, ("New Task Added"))
-        return redirect('home')
+        return redirect('todo-list')
     else:
-        all_tasks = TaskList.objects.all()
+
+        all_tasks = TaskList.objects.filter(user=request.user)
         paginator = Paginator(all_tasks, 5)
         page = request.GET.get('pg')
         all_tasks = paginator.get_page(page)
@@ -23,44 +34,58 @@ def todolist(request):
         return render(request, 'todolist.html', {"all_tasks": all_tasks})
 
 
+@login_required
 def contact(request):
     context = {}
     return render(request, 'contact.html', context)
 
 
+@login_required
 def about(request):
     context = {}
     return render(request, 'about.html', context)
 
 
+@login_required
 def delete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.delete()
-    return redirect("home")
+    if task.user == request.user:
+        task.delete()
+        return redirect("todo-list")
+    else:
+        messages.warning(request, ("Your Not Allowed!"))
+        return redirect('todo-list')
 
 
+@login_required
 def complete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.done = True
-    task.save()
-    return redirect("home")
+    if task.user == request.user:
+        task.done = True
+        task.save()
+        return redirect("todo-list")
+    else:
+        messages.warning(request, ("Your Not Allowed!"))
+        return redirect('todo-list')
 
 
+@login_required
 def pending_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
     task.done = False
     task.save()
-    return redirect("home")
+    return redirect("todo-list")
 
 
+@login_required
 def edit_task(request, task_id):
     if request.method == "POST":
         task = TaskList.objects.get(pk=task_id)
         form = TaskForm(request.POST or None, instance=task)
         if form.is_valid():
             form.save()
-        messages.success(request, ("New Task Edited!"))
-        return redirect('home')
+        messages.success(request, ("Task Edited!"))
+        return redirect('todo-list')
     else:
         task = TaskList.objects.get(pk=task_id)
         return render(request, 'edit_task.html', {"task": task})
